remove warning

+++ htdocs/bom/bom_card.php
@@ -133,7 +133,7 @@ if (empty($reshook))
 		$error = 0;
 
 		// Set if we used free entry or predefined product
-		$idprod = GETPOST('idprod', 'int');
+		$idprod = (int) GETPOST('idprod', 'int');
 		$qty = GETPOST('qty', 'int');
 		$qty_frozen = GETPOST('qty_frozen', 'int');
 		$disable_stock_change = GETPOST('disable_stock_change', 'int');
@@ -172,8 +172,7 @@ if (empty($reshook))
    			$bomline->position = ($ranktouse + 1);
 
 			$result = $bomline->create($user);
-			if ($result <= 0)
-			{
+			if ($result <= 0) {
 				setEventMessages($bomline->error, $bomline->errors, 'errors');
 				$action = '';
 			} else {
